+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-26 21:14:50
-// Commit: style(core): adjust design tokens, typography, and contrast in core
+// Updated at: 2026-04-29 09:12:27
+// Commit: fix(core): prevent double tap duplicate action in core
 
 class coreComponentState {
-  final String commitRef = '730';
-  final String lastMessage = 'style(core): adjust design tokens, typography, and contrast in core';
-  final DateTime timestamp = DateTime.parse('2026-04-26T21:14:50+05:00');
+  final String commitRef = '744';
+  final String lastMessage = 'fix(core): prevent double tap duplicate action in core';
+  final DateTime timestamp = DateTime.parse('2026-04-29T09:12:27+05:00');
 }